MediaFragment component
